Get rid of `with_multisrctop`
See the new comment in `config-ml.in`, and the email chain starting with
[1].

The short story is that I am trying to do a bunch of build system
improvements to GCC to make it easier to build GCC's runtime libraries
separately (see [1] for why). As part of this, I started looking at the
current multilib implementation, with a hunch that things could be
simpler, and with those simplifications, we could refine the idiomatic
division of labor between the outer and inner build systems in a way
that I think would be just better, and also would make the separate or
"per component" builds I want to do more maintainable.

In that linked thread I found variable/mechanism that looked unused, and
after some back and forth with Joseph Myers, we came to an agreement
that since GNU Make was mandatory now, and thus VPATH support was
assumed, there was no need to do these convoluted symlink tree things to
unnaturally allow multilib to get away with `srcdir = objdir` building.
(When indeed, `srcdir != objdir_for_machine1 != objdir_for_machine2` is
multilib in a nutshell.)

The last (as of this patch) email in the chain is Joseph writing:

> I think it's reasonable to eliminate such a workaround, but I don't know
> how the newlib community feels, should they wish to update the shared
> toplevel build infrastructure with changes from GCC.

This email is thus me submitting the same change (well, actually a
slightly improved one -- the `config-ml.in` comment is new, reflecting
what went on in the email thread) to ask you all that question.

If our reasoning looks correct, and you see no reason to support these
no-VPATH workarounds either, then we have our concensus! I think we can
land this change in both repos (GCC, newlib-cygwin), in either order (no
repo need wait on the other repo).

[1]: https://sourceware.org/pipermail/gcc-patches/2025-August/691963.html
[2]: https://blog.obsidian.systems/compiler-bootstrapping-in-nixpkgs/
[3]: https://sourceware.org/pipermail/gcc-patches/2025-August/692233.html

Signed-off-by: John Ericson <git@JohnEricson.me>